Ensure kubelet and containerd are in path for temporary startup

+++ pkg/rke2/rke2.go
@@ -2,13 +2,14 @@ package rke2
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"strings"
 	"time"
+
+	"github.com/pkg/errors"
 
 	containerdk3s "github.com/rancher/k3s/pkg/agent/containerd"
 	"github.com/rancher/rke2/pkg/controllers/cisnetworkpolicy"
@@ -20,6 +21,7 @@ import (
 	"github.com/rancher/k3s/pkg/cli/etcdsnapshot"
 	"github.com/rancher/k3s/pkg/cli/server"
 	"github.com/rancher/k3s/pkg/cluster/managed"
+	daemonconfig "github.com/rancher/k3s/pkg/daemons/config"
 	"github.com/rancher/k3s/pkg/daemons/executor"
 	"github.com/rancher/k3s/pkg/etcd"
 	rawServer "github.com/rancher/k3s/pkg/server"
@@ -27,6 +29,7 @@ import (
 	"github.com/rancher/rke2/pkg/images"
 	"github.com/rancher/rke2/pkg/podexecutor"
 	"github.com/urfave/cli"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
 )
 
@@ -42,6 +45,7 @@ const (
 	CISProfile15           = "cis-1.5"
 	CISProfile16           = "cis-1.6"
 	defaultAuditPolicyFile = "/etc/rancher/rke2/audit-policy.yaml"
+	containerdSock         = "/run/k3s/containerd/containerd.sock"
 )
 
 func Server(clx *cli.Context, cfg Config) error {
@@ -168,16 +172,30 @@ func podManifestsDir(dataDir string) string {
 	return filepath.Join(dataDir, "agent", config.DefaultPodManifestPath)
 }
 
+func binDir(dataDir string) string {
+	return filepath.Join(dataDir, "bin")
+}
+
 func removeOldPodManifests(dataDir string, disabledItems map[string]bool, clusterReset bool) error {
-	var kubeletStandAlone bool
-
-	kubeletErr := make(chan error)
-	containerdErr := make(chan error)
-
-	ctx, cancel := context.WithTimeout(context.Background(), (5 * time.Minute))
-	defer cancel()
-
+	kubeletStandAlone := false
+	execPath := binDir(dataDir)
 	manifestDir := podManifestsDir(dataDir)
+
+	// no need to clean up static pods if this is a clean install (bin or manifests dirs missing)
+	for _, path := range []string{execPath, manifestDir} {
+		if _, err := os.Stat(path); os.IsNotExist(err) {
+			return nil
+		}
+	}
+
+	// ensure etcd manifest is removed if cluster-reset is passed, and force
+	// standalone startup to ensure static pods are terminated
+	if clusterReset {
+		disabledItems["etcd"] = true
+		kubeletStandAlone = true
+	}
+
+	// check to see if there are manifests for any disabled components
 	for component, disabled := range disabledItems {
 		if disabled {
 			manifestName := filepath.Join(manifestDir, component+".yaml")
@@ -187,53 +205,55 @@ func removeOldPodManifests(dataDir string, disabledItems map[string]bool, cluste
 		}
 	}
 
-	if clusterReset {
-		// deleting old etcd if cluster reset is passed
-		manifestName := filepath.Join(manifestDir, "etcd.yaml")
-		if _, err := os.Stat(manifestName); err == nil {
-			kubeletStandAlone = true
-		}
-	}
 	if kubeletStandAlone {
 		// delete all manifests
 		for component := range disabledItems {
 			manifestName := filepath.Join(manifestDir, component+".yaml")
 			if err := os.RemoveAll(manifestName); err != nil {
-				return err
+				return errors.Wrapf(err, "unable to delete %s manifest", component)
 			}
 		}
-		kubeletCmd := exec.CommandContext(ctx, "kubelet")
-		containerdCmd := exec.CommandContext(ctx, "containerd")
 
-		tmpAddress := filepath.Join(os.TempDir(), "containerd.sock")
+		ctx, cancel := context.WithTimeout(context.Background(), (5 * time.Minute))
+		defer cancel()
+
+		kubeletCmd := exec.CommandContext(ctx, filepath.Join(execPath, "kubelet"))
+		containerdCmd := exec.CommandContext(ctx, filepath.Join(execPath, "containerd"))
+
+		kubeletErr := make(chan error)
+		containerdErr := make(chan error)
 
 		// start containerd
-		go startContainerd(dataDir, containerdErr, tmpAddress, containerdCmd)
+		go startContainerd(dataDir, containerdErr, containerdCmd)
 		// start kubelet
-		go startKubelet(dataDir, kubeletErr, tmpAddress, kubeletCmd)
-
+		go startKubelet(dataDir, kubeletErr, kubeletCmd)
 		// check for any running containers from the disabled items list
-		go checkForRunningContainers(ctx, tmpAddress, disabledItems, kubeletErr, containerdErr)
+		go checkForRunningContainers(ctx, disabledItems, kubeletErr, containerdErr)
 
-		for {
-			time.Sleep(5 * time.Second)
-			select {
-			case err := <-kubeletErr:
-				logrus.Infof("kubelet Exited: %v, exiting Containerd", err)
-				// exits containerd if kubelet exits
-				containerdCmd.Process.Kill()
+		// ensure temporary kubelet and containerd are terminated
+		defer func() {
+			if kubeletCmd.Process != nil {
 				kubeletCmd.Process.Kill()
-
-			case err := <-containerdErr:
-				logrus.Infof("Containerd Exited: %v, exiting kubelet", err)
-				kubeletCmd.Process.Kill()
-				containerdCmd.Process.Kill()
-
-			case <-ctx.Done():
-				logrus.Info("Timeout reached, exiting kubelet and containerd")
-				kubeletCmd.Process.Kill()
+			}
+			if containerdCmd.Process != nil {
 				containerdCmd.Process.Kill()
 			}
+		}()
+
+		for {
+			select {
+			case err := <-kubeletErr:
+				if err != nil {
+					return errors.Wrap(err, "temporary kubelet process exited unexpectedly")
+				}
+			case err := <-containerdErr:
+				if err != nil {
+					return errors.Wrap(err, "temporary containerd process exited unexpectedly")
+				}
+			case <-ctx.Done():
+				return errors.New("static pod cleanup timed out")
+			}
+			logrus.Info("Static pod cleanup completed successfully")
 			break
 		}
 	}
@@ -246,48 +266,57 @@ func isCISMode(clx *cli.Context) bool {
 	return profile == CISProfile15 || profile == CISProfile16
 }
 
-func startKubelet(dataDir string, errChan chan error, tmpAddress string, cmd *exec.Cmd) {
-	logrus.Infof("starting kubelet to clean up old static pods")
+func startKubelet(dataDir string, errChan chan error, cmd *exec.Cmd) {
 	args := []string{
 		"--fail-swap-on=false",
 		"--container-runtime=remote",
-		"--containerd=" + tmpAddress,
-		"--container-runtime-endpoint=unix://" + tmpAddress,
+		"--containerd=" + containerdSock,
+		"--container-runtime-endpoint=unix://" + containerdSock,
 		"--pod-manifest-path=" + podManifestsDir(dataDir),
 	}
 	cmd.Args = append(cmd.Args, args...)
-	cmd.Stdout = nil
-	cmd.Stderr = nil
+	cmd.Env = append(os.Environ(), fmt.Sprintf("PATH=%s:%s", binDir(dataDir), os.Getenv("PATH")))
+	cmd.Env = append(cmd.Env, "NOTIFY_SOCKET=")
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
 
+	logrus.Infof("Running temporary kubelet %s", daemonconfig.ArgString(cmd.Args))
 	errChan <- cmd.Run()
 }
 
-func startContainerd(dataDir string, errChan chan error, tmpAddress string, cmd *exec.Cmd) {
+func startContainerd(dataDir string, errChan chan error, cmd *exec.Cmd) {
 	args := []string{
-		"-a", tmpAddress,
+		"-c", filepath.Join(dataDir, "agent", "etc", "containerd", "config.toml"),
+		"-a", containerdSock,
+		"--state", filepath.Dir(containerdSock),
 		"--root", filepath.Join(dataDir, "agent", "containerd"),
 	}
 	cmd.Args = append(cmd.Args, args...)
-	cmd.Stdout = nil
-	cmd.Stderr = nil
+	cmd.Env = append(os.Environ(), fmt.Sprintf("PATH=%s:%s", binDir(dataDir), os.Getenv("PATH")))
+	cmd.Env = append(cmd.Env, "NOTIFY_SOCKET=")
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
 
+	logrus.Infof("Running temporary containerd %s", daemonconfig.ArgString(cmd.Args))
 	errChan <- cmd.Run()
 }
 
-func isContainerRunning(ctx context.Context, name string, resp *runtimeapi.ListContainersResponse) bool {
+func isContainerRunning(name string, resp *runtimeapi.ListContainersResponse) bool {
 	for _, c := range resp.Containers {
-		if c.Metadata.Name == name {
+		if c.Labels["io.kubernetes.pod.namespace"] == metav1.NamespaceSystem &&
+			strings.HasPrefix(c.Labels["io.kubernetes.pod.name"], name) &&
+			c.Labels["io.kubernetes.container.name"] == name {
 			return true
 		}
 	}
 	return false
 }
 
-func checkForRunningContainers(ctx context.Context, tmpAddress string, disabledItems map[string]bool, kubeletErr, containerdErr chan error) {
+func checkForRunningContainers(ctx context.Context, disabledItems map[string]bool, kubeletErr, containerdErr chan error) {
 	ticker := time.NewTicker(10 * time.Second)
 	defer ticker.Stop()
 	for range ticker.C {
-		conn, err := containerdk3s.CriConnection(ctx, tmpAddress)
+		conn, err := containerdk3s.CriConnection(ctx, containerdSock)
 		if err != nil {
 			logrus.Warnf("Failed to setup cri connection: %v", err)
 			continue
@@ -299,21 +328,22 @@ func checkForRunningContainers(ctx context.Context, tmpAddress string, disabledI
 			logrus.Warnf("Failed to list containers: %v", err)
 			continue
 		}
-		var gc bool
-		for item, _ := range disabledItems {
-			if isContainerRunning(ctx, item, resp) {
+		containersRunning := false
+		for item := range disabledItems {
+			if isContainerRunning(item, resp) {
 				logrus.Infof("Waiting for deletion of %s static pod", item)
-				gc = true
+				containersRunning = true
 				break
 			}
 		}
-		if gc {
+		if containersRunning {
 			continue
 		}
-		// if all disabled items containers has been removed
-		// then terminate kubelet and containerd
-		containerdErr <- fmt.Errorf("static pods deleted")
-		kubeletErr <- fmt.Errorf("static pods deleted")
+		// if all disabled item containers have been removed,
+		// send on the subprocess error channels to wake up the select
+		// loop and shut everything down.
+		containerdErr <- nil
+		kubeletErr <- nil
 		break
 	}
 }
